requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.49
Werkzeug==2.3.7
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
blinker==1.9.0
click==8.3.2
greenlet==3.3.2
numpy==1.26.4
Pillow==11.1.0
scipy==1.11.2
opencv-python-headless==4.10.0.84
gunicorn==21.2.0
